refactor: convert class-based views to use-cases
