+++ backend/api/urls.py
@@ -1,8 +1,6 @@
-from django.contrib import admin
-from django.urls import include, path
-
-# API Views
+from views.auth_views import signin, get_current_user
 import api.views.task as task_views
+from django.urls import path
 
 urlpatterns = [
     # Tasks API Route
@@ -10,4 +8,10 @@ urlpatterns = [
 
     # Task Detail API
     path('/tasks/<int:pk>', task_views.TaskDetail.as_view(), name='task_detail'),
+]
+
+# AUTHENTICATION URLS
+urlpatterns += [
+    path('auth/me', get_current_user, name='getme'),
+    path('auth/signin', signin, name='signin'),
 ]
