+++ backend/core/urls.py
@@ -4,9 +4,6 @@ from django.urls import include, path
 urlpatterns = [
     path('admin/', admin.site.urls),
 
-    # Authentication Routes
-    path('api/auth/', include('authentication.urls')),
-
     # API Routes
     path('api/', include('api.urls')),
 ]
